navbar and landingpage
added navigation page and worked on landing page
